APIS_AutoTest/requirements.txt
xlrd==1.2.0
XlsxWriter==1.2.8
PyMySQL==0.9.3
xlutils==2.0.0
xlwt==1.3.0
sshtunnel==0.1.5
requests==2.22.0
PyYAML==5.3.1
